Implementation of PID and Coordinate Ascent algorithm
Simple PID implementation using application framework defined
in the project files, as well as the addition of a coordinate
ascent class defined in CA.cpp(hpp).

Writeup in README.md with one simple visual showing the difference
in the controller performance for a hand tuned set of coefficients
versus the coefficients produced by the optimizer.

Note that the final performance of the controller is still quite
poor, but the optimizer ran for >200 cycles and was no longer
succesfully reducing the total error. It's possible the controller
needs another non linear term.